Initial Commit
Going through and transferring all my projects to this github account, including this.

GlennMC was a commission from May of 2020 and was done within a day so it was rushed.

Restored a lot of it and improved upon a lot of code, No plans to update it so have fun :+1: :)